Move FDR/MTC to a separate class.
